Rerfactor: move cvs2json logic to module
