+++ server.js
@@ -1,38 +1,14 @@
 const express = require('express');
+const fs = require('fs');
+const agilaSverige = require('./agilaSverige');
+
 const app = express();
 
 app.use(express.static('public'))
 
 app.get('/api/schedule', function (req, res) {
-  function toObject(row) {
-    function parseTime(s) {
-      const parts = s.split(':');
-      const now = new Date();
-      now.setHours(parts[0], parts[1], 0, 0);
-      return now;
-    }
-
-    function parseDuration(s) {
-      const parts = s.trim().split('-');
-      if (parts.length < 2) return null;
-      return {'start': parseTime(parts[0]), 'stop': parseTime(parts[1])};
-    }
-    var obj = parseDuration(row[0]);
-    if (!obj) return null;
-    obj.space = {'title': row[1], 'speaker': row[2]};
-    obj.tab =   {'title': row[3], 'speaker': row[4]};
-    return obj;
-  }
-  const fs = require('fs');
-  const csv = require('csv');
-  const JSONStream = require('JSONStream');
-
   var input = fs.createReadStream('data/2017.csv');
-  input
-    .pipe(csv.parse())
-    .pipe(csv.transform(toObject))
-    .pipe(JSONStream.stringify())
-    .pipe(res);
+  agilaSverige.cvs2json(input, res);
 });
 
 app.listen(3000, function () {
